+++ src/gei/Handlers/GenerateScriptCommandHandler.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Extensions;
 using OctoshiftCLI.GithubEnterpriseImporter.Commands;
+using OctoshiftCLI.GithubEnterpriseImporter.Services;
 using OctoshiftCLI.Handlers;
 
 [assembly: InternalsVisibleTo("OctoshiftCLI.Tests")]
@@ -21,17 +22,20 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
     private readonly GithubApi _sourceGithubApi;
     private readonly AdoApi _sourceAdoApi;
     private readonly IVersionProvider _versionProvider;
+    private readonly GhesVersionChecker _ghesVersionChecker;
 
     public GenerateScriptCommandHandler(
         OctoLogger log,
         GithubApi sourceGithubApi,
         AdoApi sourceAdoApi,
-        IVersionProvider versionProvider)
+        IVersionProvider versionProvider,
+        GhesVersionChecker ghesVersionChecker)
     {
         _log = log;
         _sourceGithubApi = sourceGithubApi;
         _sourceAdoApi = sourceAdoApi;
         _versionProvider = versionProvider;
+        _ghesVersionChecker = ghesVersionChecker;
     }
 
     public async Task Handle(GenerateScriptCommandArgs args)
@@ -188,8 +192,8 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
         }
 
         return sequential
-            ? GenerateSequentialGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, awsBucketName, awsRegion, noSslVerify, skipReleases, lockSourceRepo, downloadMigrationLogs, keepArchive)
-            : GenerateParallelGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, awsBucketName, awsRegion, noSslVerify, skipReleases, lockSourceRepo, downloadMigrationLogs, keepArchive);
+            ? await GenerateSequentialGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, awsBucketName, awsRegion, noSslVerify, skipReleases, lockSourceRepo, downloadMigrationLogs, keepArchive)
+            : await GenerateParallelGithubScript(repos, githubSourceOrg, githubTargetOrg, ghesApiUrl, awsBucketName, awsRegion, noSslVerify, skipReleases, lockSourceRepo, downloadMigrationLogs, keepArchive);
     }
 
     private async Task<string> InvokeAdo(string adoServerUrl, string adoSourceOrg, string adoTeamProject, string githubTargetOrg, bool sequential, bool downloadMigrationLogs)
@@ -250,7 +254,7 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
         return repos;
     }
 
-    private string GenerateSequentialGithubScript(IEnumerable<string> repos, string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string awsBucketName, string awsRegion, bool noSslVerify, bool skipReleases, bool lockSourceRepo, bool downloadMigrationLogs, bool keepArchive)
+    private async Task<string> GenerateSequentialGithubScript(IEnumerable<string> repos, string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string awsBucketName, string awsRegion, bool noSslVerify, bool skipReleases, bool lockSourceRepo, bool downloadMigrationLogs, bool keepArchive)
     {
         var content = new StringBuilder();
 
@@ -258,6 +262,20 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
         content.AppendLine();
         content.AppendLine(VersionComment);
         content.AppendLine(EXEC_FUNCTION_BLOCK);
+
+        content.AppendLine(VALIDATE_GH_PAT);
+        if (await _ghesVersionChecker.AreBlobCredentialsRequired(ghesApiUrl))
+        {
+            if (awsBucketName.HasValue() || awsRegion.HasValue())
+            {
+                content.AppendLine(VALIDATE_AWS_ACCESS_KEY_ID);
+                content.AppendLine(VALIDATE_AWS_SECRET_ACCESS_KEY);
+            }
+            else
+            {
+                content.AppendLine(VALIDATE_AZURE_STORAGE_CONNECTION_STRING);
+            }
+        }
 
         content.AppendLine($"# =========== Organization: {githubSourceOrg} ===========");
 
@@ -274,7 +292,7 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
         return content.ToString();
     }
 
-    private string GenerateParallelGithubScript(IEnumerable<string> repos, string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string awsBucketName, string awsRegion, bool noSslVerify, bool skipReleases, bool lockSourceRepo, bool downloadMigrationLogs, bool keepArchive)
+    private async Task<string> GenerateParallelGithubScript(IEnumerable<string> repos, string githubSourceOrg, string githubTargetOrg, string ghesApiUrl, string awsBucketName, string awsRegion, bool noSslVerify, bool skipReleases, bool lockSourceRepo, bool downloadMigrationLogs, bool keepArchive)
     {
         var content = new StringBuilder();
 
@@ -282,6 +300,20 @@ public class GenerateScriptCommandHandler : ICommandHandler<GenerateScriptComman
         content.AppendLine();
         content.AppendLine(VersionComment);
         content.AppendLine(EXEC_AND_GET_MIGRATION_ID_FUNCTION_BLOCK);
+
+        content.AppendLine(VALIDATE_GH_PAT);
+        if (await _ghesVersionChecker.AreBlobCredentialsRequired(ghesApiUrl))
+        {
+            if (awsBucketName.HasValue() || awsRegion.HasValue())
+            {
+                content.AppendLine(VALIDATE_AWS_ACCESS_KEY_ID);
+                content.AppendLine(VALIDATE_AWS_SECRET_ACCESS_KEY);
+            }
+            else
+            {
+                content.AppendLine(VALIDATE_AZURE_STORAGE_CONNECTION_STRING);
+            }
+        }
 
         content.AppendLine();
         content.AppendLine("$Succeeded = 0");
@@ -528,5 +560,33 @@ function ExecAndGetMigrationID {
         $_
     } | Select-String -Pattern ""\(ID: (.+)\)"" | ForEach-Object { $_.matches.groups[1] }
     return $MigrationID
+}";
+    private const string VALIDATE_GH_PAT = @"
+if (-not $env:GH_PAT) {
+    Write-Error ""GH_PAT environment variable must be set to a valid GitHub Personal Access Token with the appropriate scopes. For more information see https://docs.github.com/en/migrations/using-github-enterprise-importer/preparing-to-migrate-with-github-enterprise-importer/managing-access-for-github-enterprise-importer#creating-a-personal-access-token-for-github-enterprise-importer""
+    exit 1
+} else {
+    Write-Host ""GH_PAT environment variable is set and will be used to authenticate to GitHub.""
+}";
+    private const string VALIDATE_AZURE_STORAGE_CONNECTION_STRING = @"
+if (-not $env:AZURE_STORAGE_CONNECTION_STRING) {
+    Write-Error ""AZURE_STORAGE_CONNECTION_STRING environment variable must be set to a valid Azure Storage Connection String that will be used to upload the migration archive to Azure Blob Storage.""
+    exit 1
+} else {
+    Write-Host ""AZURE_STORAGE_CONNECTION_STRING environment variable is set and will be used to upload the migration archive to Azure Blob Storage.""
+}";
+    private const string VALIDATE_AWS_ACCESS_KEY_ID = @"
+if (-not $env:AWS_ACCESS_KEY_ID) {
+    Write-Error ""AWS_ACCESS_KEY_ID environment variable must be set to a valid AWS Access Key ID that will be used to upload the migration archive to AWS S3.""
+    exit 1
+} else {
+    Write-Host ""AWS_ACCESS_KEY_ID environment variable is set and will be used to upload the migration archive to AWS S3.""
+}";
+    private const string VALIDATE_AWS_SECRET_ACCESS_KEY = @"
+if (-not $env:AWS_SECRET_ACCESS_KEY) {
+    Write-Error ""AWS_SECRET_ACCESS_KEY environment variable must be set to a valid AWS Secret Access Key that will be used to upload the migration archive to AWS S3.""
+    exit 1
+} else {
+    Write-Host ""AWS_SECRET_ACCESS_KEY environment variable is set and will be used to upload the migration archive to AWS S3.""
 }";
 }