+++ src/gei/Handlers/MigrateRepoCommandHandler.cs
@@ -5,6 +5,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using OctoshiftCLI.Extensions;
 using OctoshiftCLI.GithubEnterpriseImporter.Commands;
+using OctoshiftCLI.GithubEnterpriseImporter.Services;
 using OctoshiftCLI.Handlers;
 
 namespace OctoshiftCLI.GithubEnterpriseImporter.Handlers;
@@ -19,6 +20,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
     private readonly EnvironmentVariableProvider _environmentVariableProvider;
     private readonly HttpDownloadService _httpDownloadService;
     private readonly FileSystemProvider _fileSystemProvider;
+    private readonly GhesVersionChecker _ghesVersionChecker;
     private readonly RetryPolicy _retryPolicy;
     private const int ARCHIVE_GENERATION_TIMEOUT_IN_HOURS = 10;
     private const int CHECK_STATUS_DELAY_IN_MILLISECONDS = 10000; // 10 seconds
@@ -35,6 +37,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         AwsApi awsApi,
         HttpDownloadService httpDownloadService,
         FileSystemProvider fileSystemProvider,
+        GhesVersionChecker ghesVersionChecker,
         RetryPolicy retryPolicy)
     {
         _log = log;
@@ -45,6 +48,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
         _awsApi = awsApi;
         _httpDownloadService = httpDownloadService;
         _fileSystemProvider = fileSystemProvider;
+        _ghesVersionChecker = ghesVersionChecker;
         _retryPolicy = retryPolicy;
     }
 
@@ -66,7 +70,7 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
 
         LogOptions(args);
 
-        var blobCredentialsRequired = await DetermineIfBlobCredentialsRequired(args);
+        var blobCredentialsRequired = await _ghesVersionChecker.AreBlobCredentialsRequired(args.GhesApiUrl);
 
         ValidateOptions(args, blobCredentialsRequired);
 
@@ -313,35 +317,6 @@ public class MigrateRepoCommandHandler : ICommandHandler<MigrateRepoCommandArgs>
             await Task.Delay(CHECK_STATUS_DELAY_IN_MILLISECONDS);
         }
         throw new TimeoutException($"Archive generation timed out after {ARCHIVE_GENERATION_TIMEOUT_IN_HOURS} hours");
-    }
-
-    private async Task<bool> DetermineIfBlobCredentialsRequired(MigrateRepoCommandArgs args)
-    {
-        var blobCredentialsRequired = false;
-
-        if (args.GhesApiUrl.HasValue())
-        {
-            blobCredentialsRequired = true;
-
-            _log.LogInformation("Using GitHub Enterprise Server - verifying server version");
-            var ghesVersion = await _sourceGithubApi.GetEnterpriseServerVersion();
-
-            if (ghesVersion != null)
-            {
-                _log.LogInformation($"GitHub Enterprise Server version {ghesVersion} detected");
-
-                if (Version.TryParse(ghesVersion, out var parsedVersion))
-                {
-                    blobCredentialsRequired = parsedVersion < new Version(3, 8, 0);
-                }
-                else
-                {
-                    _log.LogInformation($"Unable to parse the version number, defaulting to using CLI for blob storage uploads");
-                }
-            }
-        }
-
-        return blobCredentialsRequired;
     }
 
     private string GetGithubRepoUrl(string org, string repo, string baseUrl) => $"{baseUrl ?? DEFAULT_GITHUB_BASE_URL}/{org}/{repo}".Replace(" ", "%20");