+++ src/gei/Program.cs
@@ -8,6 +8,7 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
 using OctoshiftCLI.Contracts;
 using OctoshiftCLI.Extensions;
+using OctoshiftCLI.GithubEnterpriseImporter.Factories;
 
 namespace OctoshiftCLI.GithubEnterpriseImporter
 {
@@ -39,6 +40,7 @@ namespace OctoshiftCLI.GithubEnterpriseImporter
                 .AddSingleton<ISourceGithubApiFactory>(sp => sp.GetRequiredService<GithubApiFactory>())
                 .AddSingleton<SecretScanningAlertServiceFactory>()
                 .AddSingleton<CodeScanningAlertServiceFactory>()
+                .AddSingleton<GhesVersionCheckerFactory>()
                 .AddHttpClient("NoSSL")
                 .ConfigurePrimaryHttpMessageHandler(() => new HttpClientHandler()
                 {